Barcode - Commented Unit test for now. pgm works.
