+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"google_mlkit_barcode_scanning","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\google_mlkit_barcode_scanning-0.3.0\\\\","native_build":true,"dependencies":["google_mlkit_commons"]},{"name":"google_mlkit_commons","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\google_mlkit_commons-0.2.0\\\\","native_build":true,"dependencies":[]},{"name":"image_picker_ios","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\image_picker_ios-0.8.5+5\\\\","native_build":true,"dependencies":[]},{"name":"path_provider_ios","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_ios-2.0.10\\\\","native_build":true,"dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\flutter_plugin_android_lifecycle-2.0.6\\\\","native_build":true,"dependencies":[]},{"name":"google_mlkit_barcode_scanning","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\google_mlkit_barcode_scanning-0.3.0\\\\","native_build":true,"dependencies":["google_mlkit_commons"]},{"name":"google_mlkit_commons","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\google_mlkit_commons-0.2.0\\\\","native_build":true,"dependencies":[]},{"name":"image_picker_android","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\image_picker_android-0.8.5\\\\","native_build":true,"dependencies":["flutter_plugin_android_lifecycle"]},{"name":"path_provider_android","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_android-2.0.16\\\\","native_build":true,"dependencies":[]}],"macos":[{"name":"path_provider_macos","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_macos-2.0.6\\\\","native_build":true,"dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_linux-2.1.7\\\\","native_build":false,"dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_windows-2.1.0\\\\","native_build":false,"dependencies":[]}],"web":[{"name":"image_picker_for_web","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\image_picker_for_web-2.1.8\\\\","dependencies":[]}]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"google_mlkit_barcode_scanning","dependencies":["google_mlkit_commons"]},{"name":"google_mlkit_commons","dependencies":[]},{"name":"image_picker","dependencies":["image_picker_android","image_picker_for_web","image_picker_ios"]},{"name":"image_picker_android","dependencies":["flutter_plugin_android_lifecycle"]},{"name":"image_picker_for_web","dependencies":[]},{"name":"image_picker_ios","dependencies":[]},{"name":"path_provider","dependencies":["path_provider_android","path_provider_ios","path_provider_linux","path_provider_macos","path_provider_windows"]},{"name":"path_provider_android","dependencies":[]},{"name":"path_provider_ios","dependencies":[]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]},{"name":"path_provider_windows","dependencies":[]}],"date_created":"2022-07-06 19:12:38.695267","version":"3.0.1"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"google_mlkit_barcode_scanning","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\google_mlkit_barcode_scanning-0.3.0\\\\","native_build":true,"dependencies":["google_mlkit_commons"]},{"name":"google_mlkit_commons","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\google_mlkit_commons-0.2.0\\\\","native_build":true,"dependencies":[]},{"name":"image_picker_ios","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\image_picker_ios-0.8.5+5\\\\","native_build":true,"dependencies":[]},{"name":"path_provider_ios","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_ios-2.0.10\\\\","native_build":true,"dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\flutter_plugin_android_lifecycle-2.0.6\\\\","native_build":true,"dependencies":[]},{"name":"google_mlkit_barcode_scanning","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\google_mlkit_barcode_scanning-0.3.0\\\\","native_build":true,"dependencies":["google_mlkit_commons"]},{"name":"google_mlkit_commons","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\google_mlkit_commons-0.2.0\\\\","native_build":true,"dependencies":[]},{"name":"image_picker_android","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\image_picker_android-0.8.5\\\\","native_build":true,"dependencies":["flutter_plugin_android_lifecycle"]},{"name":"path_provider_android","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_android-2.0.16\\\\","native_build":true,"dependencies":[]}],"macos":[{"name":"path_provider_macos","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_macos-2.0.6\\\\","native_build":true,"dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_linux-2.1.7\\\\","native_build":false,"dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_windows-2.1.0\\\\","native_build":false,"dependencies":[]}],"web":[{"name":"image_picker_for_web","path":"C:\\\\SDK\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\image_picker_for_web-2.1.8\\\\","dependencies":[]}]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"google_mlkit_barcode_scanning","dependencies":["google_mlkit_commons"]},{"name":"google_mlkit_commons","dependencies":[]},{"name":"image_picker","dependencies":["image_picker_android","image_picker_for_web","image_picker_ios"]},{"name":"image_picker_android","dependencies":["flutter_plugin_android_lifecycle"]},{"name":"image_picker_for_web","dependencies":[]},{"name":"image_picker_ios","dependencies":[]},{"name":"path_provider","dependencies":["path_provider_android","path_provider_ios","path_provider_linux","path_provider_macos","path_provider_windows"]},{"name":"path_provider_android","dependencies":[]},{"name":"path_provider_ios","dependencies":[]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]},{"name":"path_provider_windows","dependencies":[]}],"date_created":"2022-07-08 09:42:05.570185","version":"3.0.1"}
