+++ ios/Flutter/Generated.xcconfig
@@ -1,8 +1,8 @@
 // This is a generated file; do not edit or check into version control.
-FLUTTER_ROOT=/Users/Akila/flutter
-FLUTTER_APPLICATION_PATH=/Users/Akila/Documents/GitHub/summer2022
+FLUTTER_ROOT=C:\SDK\flutter
+FLUTTER_APPLICATION_PATH=C:\Users\sdeje\Downloads\barcode_feature
 COCOAPODS_PARALLEL_CODE_SIGN=true
-FLUTTER_TARGET=lib/main.dart
+FLUTTER_TARGET=lib\main.dart
 FLUTTER_BUILD_DIR=build
 FLUTTER_BUILD_NAME=1.0.0
 FLUTTER_BUILD_NUMBER=7
