+++ ios/Flutter/flutter_export_environment.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
 export "FLUTTER_ROOT=C:\SDK\flutter"
-export "FLUTTER_APPLICATION_PATH=C:\Users\sdeje\OneDrive\Desktop\summer2022"
+export "FLUTTER_APPLICATION_PATH=C:\Users\sdeje\Downloads\barcode_feature"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_TARGET=lib\main.dart"
 export "FLUTTER_BUILD_DIR=build"
