+++ macos/Flutter/ephemeral/Flutter-Generated.xcconfig
@@ -1,6 +1,6 @@
 // This is a generated file; do not edit or check into version control.
 FLUTTER_ROOT=C:\SDK\flutter
-FLUTTER_APPLICATION_PATH=C:\Users\sdeje\OneDrive\Desktop\summer2022
+FLUTTER_APPLICATION_PATH=C:\Users\sdeje\Downloads\barcode_feature
 COCOAPODS_PARALLEL_CODE_SIGN=true
 FLUTTER_BUILD_DIR=build
 FLUTTER_BUILD_NAME=1.0.0
