+++ pubspec.yaml
@@ -34,7 +34,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  google_mlkit_barcode_scanning: ^0.3.0
+  google_mlkit_barcode_scanning: any #^0.3.0
   image_picker: ^0.8.5+3
   googleapis_auth: ^1.3.1
   googleapis: ^9.1.0
