+++ test/barcode_scanner_test.dart
